David V1.0
Minimal changes. So minor I will not rename this version as there were
no system/game changes. Cleaned up unnecessary global variables a bit in
displaypanel. Minor change to name of the class so that rhythm is
actually spelled correctly. Added comments to the code so it hopefully
is easier to understand. Nothing else done yet, all bugs previously
mentioned are still in the current version.