V2.0
The is currently in the process of being ported to the slick2d graphics
library. The old version of the game, based off the built in swing
graphics library was found not be compatible for the purposes that we
want to use it for. Very little has been done, just a shell has been
outlined, but hopefully people will be able to see the basic structure
of the game. Bug list was also updated to show that all bugs so far
found are obsolete with the new port.