V2.1
Game has been ported to slick. All work previously done is basically
replicated in slick. A new idea was formed and now there are 5 "radar"
circles, as I am calling them, which will correspond to different keys
on the keyboard. A rhythm circle will appear on one of the many circles
when the selector is 180 degrees away from the rhythm circle. The logic
to determine this was loosely implemented. Sample music has been added
and plays when the game is started. A debug mode was also added to
determine the time between updates and renders so that it would be
easier to control the speed of the selector. To activate debug mode in
game press `.