V2.3
Circles now take 3 arguments: an angle (which is the location that they
will display on the screen), a rotation (which is the number of times
the selector has gone around before the circle is displayed), and a
layer (which is the radar circle which the circle lies on [1-5]).
Circles also disappear after the selector has passed by them now, though
smoothly disappearing would be nicer. Finally the circles are now being
read from a textfile located in data and the sample music that is
normally playing is no longer playing as the code that plays it has been
commented out for testing purposes.

Found Bugs:
None.
Fixed Bugs:
None.