Shady Rhythms V2.4
The final core part of the game is here. Collision detection was added
based on angle which is selector is at. Points were also added so that
you gain points based on how close you are to the center of the rhythm
circle when you hit the key. The next thing to do is to create a menu
system and add the storyline into the game.

Found Bugs:
0_0_5
Check bug list for more info.