Rhythm Generator V1.0
As you may have seen from last commit, naming conventions have changed.
We now have a rhythm generator, which we can use to generate rhythms.
Commits will come in two types. One will be the main game, which will
only give changes to the main game and whose version name will be
preceded by the name of the game (working name is Shady Rhythms). The
second type of commit is updates to the rhythm generator which will have
the version preceded by the name Rhythm Generator.
This is the first version of the rhythm generator. It will play songs
with the basic structure of our game and when you hit keys the rhythms
will be written to a text file. Working keys for the circles are 'j',
'k', 'l', ';', ' '.  In order to finish writing to a text file, when
done with the rhythm press the escape key. Then press alt-F4 to quit the
game and copy the song.txt into the data folder of the main game.